Link players to the games they have played
Games must keep track of who is `player_one` (white), and who is
`player_two` (black). This means that the all the games that a player
has been involved in are the combination of all the games in which they
were `player_one`, and all the games in which they were `player_two`.

This commit adds `Player#games`, to combine these two Relations in order
to get all of their games.