Change `.container` to flexbox
This makes it easier to keep a sticky footer while still centering the
elements. Doing it the css-tricks way without having the container flex
its contents was causing a weird vertical scrollbar to pop up. I will
postop why this was the case later on. Greg could not figure it out
either.

Doing it this way:
https://css-tricks.com/couple-takes-sticky-footer/#article-header-id-0